An 1249 lp actions (#25)
* LP actions model and tests

* Fix to one token LPs

* flattened out tokens, added decimal, updated docs and test

* pool id change - remove array

* aggregates currencies and groups by msg_index

* Revert "pool id change - remove array"

This reverts commit dc4185d51f739224d5d90745d3e6cf4f126d0d58.

* lucid chart change round 1

* transfers updates

* LP actions - flattened

* change vote weight to float

+++ models/silver/silver__governance_votes.sql
@@ -21,7 +21,7 @@ WITH vote_options AS (
         ELSE
             TRY_PARSE_JSON(attribute_value):option 
       END AS vote_option, 
-      TRY_PARSE_JSON(attribute_value):weight :: INTEGER AS vote_weight
+      TRY_PARSE_JSON(attribute_value):weight :: FLOAT AS vote_weight
   FROM {{ ref('silver__msg_attributes') }}
   WHERE msg_type = 'proposal_vote'
   AND attribute_key = 'option'
